+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Note that the `[Unreleased]` section contains all changes that haven't yet made 
 - Added `--no-build` and `--no-check` options (#211).
 - Added extension support (#212).
 - Format makedeb-styled message calls (`msg`, `msg2`, `warning`, etc.) in PKGBUILD functions with a prefix (#212).
+- Added support for the `NO_COLOR` environment variable (#188).
 
 ### Removed
 - Dropped support for VCS systems besides Git and SVN (#210).